Begin migration to SDL3
Also switch to ImGui v1.92.6

+++ src/RealEngine/graphics/fonts/RasterizedFont.hpp
@@ -14,7 +14,7 @@ namespace re {
 
 struct RasterizedFontCreateInfo {
     std::span<const unsigned char> ttfBytes{};
-    int pointSize{};
+    float pointSize{};
     int faceIndex{};
     std::span<const UnicodeRange> ranges = k_asciiPrintableUnicodeRanges;
 };
